refactor: Initial re-org of examples.
