+++ examples/createPool/v2/createPoolComposableStable.ts
