+++ examples/createPool/v2/createPoolWeighted.ts
