+++ examples/createPool/v3/createPoolWeighted.ts
