+++ examples/lib/executeExample.ts
@@ -1,7 +1,7 @@
-import { stopAnvilForks } from '../test/anvil/anvil-global-setup';
+import { stopAnvilForks } from '../../test/anvil/anvil-global-setup';
 
 const executeExample = async (filePath) => {
-    const exampleModule = await import(`../${filePath}`);
+    const exampleModule = await import(`../../${filePath}`);
     const example = exampleModule.default;
     await example();
     stopAnvilForks();
