+++ examples/removeLiquidity/removeLiquidityRecovery.ts
@@ -1,9 +1,9 @@
 /**
- * Example showing how to remove liquidity from a pool.
+ * Example showing how to remove liquidity from a pool in recovery mode.
  * (Runs against a local Anvil fork)
  *
  * Run with:
- * pnpm example ./examples/removeLiquidity.ts
+ * pnpm example ./examples/removeLiquidityRecovery.ts
  */
 import dotenv from 'dotenv';
 dotenv.config();
@@ -15,22 +15,20 @@ import {
     ChainId,
     InputAmount,
     PoolState,
-    PriceImpact,
     RemoveLiquidity,
-    RemoveLiquidityInput,
     RemoveLiquidityKind,
+    RemoveLiquidityRecoveryInput,
     Slippage,
-} from '../src';
-import { ANVIL_NETWORKS, startFork } from '../test/anvil/anvil-global-setup';
-import { makeForkTx } from './utils/makeForkTx';
+} from '../../src';
+import { ANVIL_NETWORKS, startFork } from '../../test/anvil/anvil-global-setup';
+import { makeForkTx } from '../lib/makeForkTx';
 
 const removeLiquidity = async () => {
     // User defined:
     const chainId = ChainId.MAINNET;
     const userAccount = '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045';
     const poolId =
-        '0x5c6ee304399dbdb9c8ef030ab642b10820db8f56000200000000000000000014'; // 80BAL-20WETH
-    const tokenOut = '0xba100000625a3754423978a60c9317c58a424e3D'; // BAL
+        '0x156c02f3f7fef64a3a9d80ccf7085f23cce91d76000000000000000000000570'; // vETH/WETH
     const slippage = Slippage.fromPercentage('1'); // 1%
 
     // Start a local anvil fork that will be used to query/tx against
@@ -49,25 +47,19 @@ const removeLiquidity = async () => {
         decimals: 18,
         address: poolState.address,
     };
-    const removeLiquidityInput: RemoveLiquidityInput = {
+    const removeLiquidityRecoveryInput: RemoveLiquidityRecoveryInput = {
         chainId,
         rpcUrl,
         bptIn,
-        tokenOut,
-        kind: RemoveLiquidityKind.SingleTokenExactIn,
+        kind: RemoveLiquidityKind.Recovery,
     };
 
-    // Calculate price impact to ensure it's acceptable
-    const priceImpact = await PriceImpact.removeLiquidity(
-        removeLiquidityInput,
-        poolState,
-    );
-    console.log(`\nPrice Impact: ${priceImpact.percentage.toFixed(2)}%`);
+    // No need to calculate Price Impact as it's always zero for remove liquidity recovery
 
     // Simulate removing liquidity to get the tokens out
     const removeLiquidity = new RemoveLiquidity();
-    const queryOutput = await removeLiquidity.query(
-        removeLiquidityInput,
+    const queryOutput = await removeLiquidity.queryRemoveLiquidityRecovery(
+        removeLiquidityRecoveryInput,
         poolState,
     );
